test/getdents: add getdents test
Add a test case for io_uring getdents. It first creates a directory and
then creates N empty files in it. File names are pure numbers from 0 to
N-1. After that, we use io_uring getdents to traverse all the items in
this directory and test if all the files are visited exactly once.

Signed-off-by: Hao Xu <howeyxu@tencent.com>